scripts/requirements.txt
autopep8==2.0.2
numpy==1.25.2
pandas==2.0.3
pycodestyle==2.11.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
tzdata==2023.3
